TM - bug 13561 - GTS.BOUNDEDPIVLs now do not produce xml containing specializationType and xsiType of rtheir interval's inner elements. This was necessary as the xsiTypes for these elements are completely different than for a standard interval (one not contained in a bounded pivl), and, according to Lloyd, completely unnecessary in their context.
This solution is a variant of a solution I put in place a while back that did this change for all intervals, not just bounded pivls. I reverted that change as I was concerned that there could be intervals that required these attributes. Now the solution is only for bounded pivls, not all intervals.